requirements.txt
pymongo
streamlit
together
Pillow
requests
langchain
langchain-core
fastembed
sentence-transformers
